requirements.txt
streamlit
google-generativeai
langchain-google-genai
faiss-cpu
sentence-transformers
numpy
python-dotenv
